Move `CommitForm` below uncommitted files
